Update UI documentation and user page template
Added screenshots to the documentation to illustrate UI improvements for posts and comments. Updated docs to describe changes based on client feedback, including the addition of delete icons and readable dates. Fixed a link in the user page template to use the correct post reference. Added a note in README about default account passwords.